+++ hosts/default/config.nix
@@ -117,7 +117,10 @@
   networking.timeServers = options.networking.timeServers.default ++ [ "pool.ntp.org" ];
 
   # Set your time zone.
-  time.timeZone = "Asia/Seoul";
+  services.automatic-timezoned.enable = true; #based on IP location
+  
+  #https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+  #time.timeZone = "Asia/Seoul"; # Set local timezone
 
   # Select internationalisation properties.
   i18n.defaultLocale = "en_US.UTF-8";